[sil-verifier] Convert else-if chain to if + continue to clarify control flow.
else-if chains provide the possibility that there is additional code after the
else-if chain. In this case, no such code exists, so by using early continues we
communicate that the iteration ends at the end of an if block and the programmer
does not need to read to the end of the else-if block.

This also allows for me to eliminate one level of indentation in the successor
code as well.